🐛 Don't try and use the solid name, use the function name instead.
the solid name ends up with things like foo_2, foo_3, etc when a solid is instantiated multiple times in a pipeline. `context.solid.definition.name` is more immune to that but it still gets the new name with a configured() which isn't what we want. I don't see any way to find the original definition through a configured() so just revert back to a simpler source of truth for now I guess.